Changed to nedb
MongoDB was working fine, but it had to be running on a serier, which is not what is needed for the app.  Eventually, I got nedb loading and not deleting the records from my json file.  However, the records are not showing in my search page.  I don't think it's actually loading properly, but I'll look into this in more detail.  Images still have to be included.