CCan;t ge search wokinng
I've tried loads of stuff, but nothing seems to work.  I've moved away from this for now and I'm working on inserting the image into the searchpage.  This is kind of working, in that, I'm getting the image icon (like when an image url isn't loading on a website), but it's closer than I've been all day!