Everything is going wrong
I worked on trying to insert the images.  This was originally just giving back the image path, but presently it's like it can't find the image from the url. I'm really stuck on this.  I moved on from that to try to get jquery working, but it' like the app is not rendering my custom.js page.  It seems to be getting it, but not doing anything with it.  I've searched for similar issues, and I found some that said to use window.$ = window.jQuery = require('./javascripts/jquery-1.12.1.min.js');, which I am doing, but window is not defined.  I also tried var jQuery = require('jquery'), which I also found, but this is giving me another error.  I've tried to initialise the searching of the database, but like everything else right now, this isn't working either.  I seem to be going round in circles since last night!