Some stuff working, lots not working
I have the records from the database showing in a list on the search page, which is great, exactly what I wanted, but I also want to be able to search the database, which is just not happening at the moment. The console is logging out the search criteria I enter, but I'm getting an error.  It's also very difficult to use jquery, which I need, with electron.  I've found a little help on this, but it's very unclear.  I still can't get the images in the directory to appear on the page.  I have the image path in the db, but I'm still getting the icon to show that there is an image there, but it's like it can't be found.  I feel that I'm going around in circles to try to fix the problems and getting nowhere!